projects/api-testing-framework/requirements.txt
# API Testing Framework Dependencies

# Core
python-dotenv>=1.0.0

# LLM API Clients
google-generativeai>=0.3.0  # Gemini
openai>=1.0.0               # OpenAI (GPT-4, GPT-3.5)
anthropic>=0.18.0           # Anthropic (Claude)

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
